Remove Optimize define (#391)
Defining Optimize in the project file is not precisely the same as specifying -c Release. (For example, DEBUG is defined instead of RELEASE, so any code which has tracing will include that tracing.) Instead of continuing to chase down all the ways it is different I think just pulling this out and requiring users to pass -c Release is a good path.