Add new automation support to runner
We need to set the cli commit hash and commit date to the environment
before we run the benchmarks for the new reporting tool. I have taken
the code from where we do that as part of the sending to benchview and
put it right before we run the benchmarks. I also make sure that this is
reset after we run.