Add python script to generate the HelixPreCommands script that will set up the environment (#512)
Adds ci_setup.py which is to be run on the build machines in the lab. This script:

* downloads dotnet
* creates a machine setup script that is to be run as part of the helixprecommands which will set up the environment. 

This change also allows us to specify all the cli arguments and use what was supplied to pass to bench view to keep that running. 

Also sets up the yaml  to use this script.