Generalize scripting to prepare for cross-repo use (#576)
* Generalize scripting to prepare for cross-repo use

This change does the following:

1. Adds a --get-perf-hash, which will get the hash of the performance
repo where the script running resides. This will be used when running in
the product repos. This will cause us to run git rev-parse HEAD to get
the hash.
2. Changes the config string to not add quotes for windows (quotes are
needed in bash, but not cmd. Currently, windows configs are being added
as key: "CompilationMode; value: Tiered". We'll need to fix it in the
database.
3. Have the run method in common print out the output of the process
that is run. We need this so that we can get the hash when running git
rev-parse HEAD.