Disable SourceLink (#608)
The test machines we run on do not have git, which SourceLink looks for.
This can cause us to get an error during build, since it tries to
Path.Combine the path to git (null) with another folder, throwing an
exception. We do not need SourceLink for the perf repo, so explicitly
disable it.