Fix Arm64 runs
The ci setup script was trying to run an arm64 dotnet in a cross
container, which was causing a failure, because the container is not,
itself arm64. The workaround is to run the ci setup script in a normal
container, pull down the x64 dotnet to get the information out of it,
but then on the helix machine, pull down the correct dotnet. Since we
already remove dotnet on linux machines, we have to download dotnet again
on the helix machines, so this change only affects the dotnet on the
build machine.