Do not install dotnet with -Channel if versions are specified (#634)
When we supply a version, but still pull down with -Channel, we will use
whichever sdk is newer. So if we are trying to check an older version,
or if there is a new version between when we start a run and when we actually
run, we will be testing the "wrong" version, ie, not the version we specified.