Move the download of the benchview scripts to ci_setup
With the influx of new runs, we have started seeing 403 errors when
downloading the benchview scripts because we are hitting the server too
frequently. Since we do not need to download the scripts in every
partition, just once on the build machine, we should move the download
to the ci_setup script. This change does that, and adds a
skip-download-benchview flag to be used in the lab so we don't try to
download them again.

Another option would be to not delete the benchview folder if it exists
already and not attempt to download the scripts if they are already on
the machine, but I did not go that route because I understand that if
there are updates to the scripts, we would want to overwrite them (on
personal machines, I guess). A third option would have been to finally
just check the benchview scripts into the performance repo, but I
decided against that since benchview will only be in use until preview7
ships.