Add System.Text.Json to the serialization tests for comparison.
Note: this doesn't full work. There is an exception in the runtime of these tests where the serialization of Dictionary<int, string> isn't a valid thing for System.Text.Json to handle as of 3.0 preview 7. The rest seems to work well and...I hope I added these correctly. We're trying to get an apples to apples comparison of serializers we can use in .NET Core for ASP.NET Core and async streams. We figured adding them here for everyone is the best course of action.