Add XUnit-style xml output for ResultsComparer
Running in AzDO, helix has support for parsing xml output from test legs
if the xml is in a particular format. This change adds the --xml option
which enables us to output these xunit-style xml output files that we
can use when running ResultsComparer in the lab to compare two
performance runs.