Update Azure SDK libs to RTM
The release of the new Azure SDK libs broke our upload because of the
way that the dependencies for the azure.core package were specified.
This fix moves us forward to the "1.0" of the new Azure SDKs and fixes
some breaking changes involved with that move. This also removes the
uploading to a queue as we are moving to another upload solution.