Fix command line matching
UseShellExecute influences the format of the command line in the ETW event.

We have to use UseShellExecute=true for GUI apps (so we can set WindowStyles) so we'll fix what we're passing in.

Also add a trim to remove a weird space that's coming from somewhere I can't figure out.