Support .NET desktop in GCPerfSim
Modify GCPerfSim to support building in desktop .NET.
It can be used with the infra using a coreclr of `self_contained: true`.

Another miscellaneous change is to make `cpu_group_number` truly optional;
before `NumaNodeInfo` could be written correctly on linux but not parsed back.

Other formatting changes are due to running the linter.