Make reporting more explicit
A recent change to the reporting library changed the semantic of how tools should detect whether or not they should try and write results. I missed updating the exporter in the microbenchmarks project. It is only broken when you try to use more than one runtime at once.

Fix is to just expose a property with the information. I don't want to bake it just into the logic of this class because different tools have different ideas of whether or not to use this functionality at all.