Update run strategy
In order to run the 5.0.1xx branch of installer we need to have darc update
our Nuget.config to contain the paths to the binaries that match the version.
As part of this we can no longer have a dependency on the master branch of
installer. So this change moves all of the master runs over to a scheduled
kickoff and leaves the CI runs to only by the release branches.