Add dotnet watch with hot reload scenario (#1759)
* Add dotnet watch with hot reload scenario

This change adds dotnet watch with hot reload as a parsable scenario for startup. I
have added new capabilties to startup in order to allow this. I have refactored the
call to `RunIteration` to allow for all of the steps that can happen in any startup
test that we have. I have also added new steps for the dotnet watch, console
listeners, that can be used to understand when a process had made certain progress.
I have also added the ability to run a process such that we return from our call,
and the underlying process can still be executing.

* Remove debug prints

* Fix up InnerLoop parser call

* Address PR Feedback

* This changes the funcs passed to `RunIteration` to also take a string param
that is the string to find in the output.
* This also creates an `IPRocessHelper` interface that is implemented by two
types. `ManagedProcessHelper` is the old `ProcessHelper` type and `RawProcessHelper`
is the new type that does not handle shutdown and also does not have default
stdout, stderr processing.

* Add handling for stderr

* Update src/tools/ScenarioMeasurement/Util/ProcessHelper.cs

Co-authored-by: Bill Wert <billwert@microsoft.com>