Changed YML PR trigger to include only release/3.1.4xx
Changed the PR trigger to be based off of release/3.1.4xx branch instead of master. This should cause all tests to trigger for dotnet maestro bot PR's. This is the minimum change to get maestro auto-merge working. Other main option would be to add specific tests that only run for when maestro does a PR.