Removed non-main pipeline jobs from pipeline
Removed non-main branch pipeline jobs from YML file. This is a part of the larger segmenting of the different release pipelines so that they exist individuallyenabling the status badge scenario and making the release that individual pipelines ran for more clear. This also allows each release to have its own jobs run and cuts down on the time it takes for an individual run for a release takes.