Removed non-5.0.1xx jobs from pipeline. (#1953)
* Removed non-5.0.1xx jobs from pipeline.

Segmented 5.0.1xx channel from the other channels so only release/5.0.1xx will run in this repo. This is a part of a trifecta of PRs on this branch, 3.1.4, and the main branch.

* Removed scheduled and manual run definitions

Removed the scheduled and manual run definitions because we only want to run new code and we don't want them to just sit around commented for clarity that they are not needed.

* Updated the yml file based on main

Updated the yml file with the updates that had been done on main but not backported. This mostly consisted of some scenario title renaming, the addition of the machine pool, and the modification of the current vm version being used.

* Updated BenchmarkDotNet reference version

Updated the BenchmarkDotNet reference version to match what is currently in the main branch. This should fix the Windows errors occuring.

* Added Nuget Source for BenchmarkDotNet

* Fix OS Ambiguity errors.

* Updated to the new job queue.