Add `--dotnet-path` to the ci scripts
For wasm builds, we use a custom dotnet populated with the bits from a
`runtime` build. And this is passed to bdn via `--cli` argument.

The project is built by `benchmarks_ci.py` first, and then run, which
then runs builds the project again via BDN, using the dotnet from the
`--cli` argument.

But this means that the project gets built with the "default" dotnet
first, and then the custom one next, which can result in a corrupted
build.