blazor_scenarios: Fix issues caused by `dotnet` being shared by all the workitems (#2367)
* blazor_scenarios: Add new `--readonly-dotnet` arg to pre, and post

.. commands, when the dotnet installation is shared, and should not be
modified. Workload install/uninstall is one such modification.

This is useful for runtime runs, where the dotnet used already has the
workload installed, and is shared by all the work items via the helix
correlation payload, and thus should not be modified.

* blazor_scenarios: Ensure that native relinking is used only when

.. intended. If the wasm-tools workload is installed, then relinking is
the default when publishing. To ensure that we don't relink, when
building with the workload installed, we can use
`WasmNativeWorkload=false`, which essentially disables the workload.

This will help enforce the existing behavior:
`blazorminapp` - uses relinking
`blazorpizza` - does not use relinking

* fix build

* fix cleanup

* revert clean_directories change

* Address review feedback