Add PaintDotNet startup tests
This change adds the PaintDotNet startup tests to the scenarios runs for the performance repo.
It also adds the scenarios run for Arm64 across all tests. In order to accomplish this changes
were made to the downloading of the dotnet SDK to enable the use of an Arm64 SDK on Arm64 runs.