Set `$(RollForward)` in MicroBenchmarks.csproj conditionally for wasm
The project is built for the host platform, like linux-x64. So, we need
to introduce a new property that would indicate that we are building for
use with wasm. And that would allow setting `$(RollForward)=LatestMajor`
in the csproj, conditionally.

For this, we pass `/p:BuildingForWasm=true` to the build command.

`dotnet.py`'s `build` function has `*args` as the last argument. And the
call site uses keyword arguments. I couldn't figure out how to pass
`args` with rest of the arguments using keywords.