Basic linker benchmarks (#2581)
Adds 2 linker benchmarks that run with TrimMode=Copy and TrimmmerDefaultAction=Copy to analyze all code in a net7.0 hello world app. BasicBenchmark reflects over the linker assembly to get the Main method of the linker and run with command line arguments. MSBuildBenchmark shells out to MSBuild to publish and force the linker to run on the project.
The tests will use a darc subscription to get the latest version of the linker tasks assembly.