dotnet.py: `--output` is no longer supported with `dotnet build/publish`
.. with .sln files. Instead, pass `/p:PublishDir=..` to the build
command.

This changed in https://github.com/dotnet/sdk/pull/29065, and broke
`dotnet-runtime-perf` pipeline's blazor scenarios.

```
$ dotnet publish /home/helixbot/work/A8850905/p/performance/src/scenarios/blazorpizza/app/BlazingPizza.sln --configuration Release --output pub /p:NuGetPackageRoot=/home/helixbot/work/A8850905/w/A9A608EA/u/artifacts/packages/ /p:UseSharedCompilation=false /p:BuildInParallel=false /m:1 --framework net7.0 /p:_TrimmerDumpDependencies=true -bl:./traces/blazor_publish.binlog -p:WasmNativeWorkload=false
MSBuild version 17.5.0-preview-22601-03+a2490dd3f for .NET

/home/helixbot/work/A8850905/p/dotnet/sdk/8.0.100-alpha.1.22606.3/Current/SolutionFile/ImportAfter/Microsoft.NET.Sdk.Solution.targets(36,5):
    error NETSDK1194: The "--output" option isn't supported when building a solution. [/home/helixbot/work/A8850905/p/performance/src/scenarios/blazorpizza/app/BlazingPizza.sln]
```

Details:

When building a solution, passing a relative path to `PublishDir` gets
evaluated per-project. So, if we pass `-p:PublishDir=pub` then we get
`pub/` sub-directories for each of the projects.

But when used with `--output pub`, output for all the projects goes to
the same directory. To have the same behavior use an absolute path.