Add Maui Mobile Testing to the perf repo (port from release/7.0 testing).
Move maui android scenario yml template to the correct place. (Private runs)

specify that it is the dotnet packs we are removing from the correlation staging

Testing the passing of links for the dotnet version.

Add rollback file for installing the maui workload on .Net versions 8 and up, and update output dir functionality to work when it is not included.