[Release/7.0] Add Maui Mobile Android Benchmarks (#2789)
Add Maui Mobile Android Benchmarks to the dotnet performance runs, rather than having them in the dotnet runtime runs. This included making new scenario folders for each and moving the logic for running and building the apps, along with changes for keeping track of versions and commit date time gathering.