Add Android BenchmarkDotnet Scenario (#2955)
Implemented an AndroidInstrumentation class for running android instrumentation tests. Added the androidBDN scenario to support the pre, test, and post.py making use of this new run type. Made the PerfLabExporter public to allow BDN runs not using the recommended config to be able to use the exporter (specifically android in this case).