Setup android power consumption test capability (#3109)
Setup the PowerConsumption tool skeleton with an implementation for the AndroidPowerParser. Also setup the python wrapper to build the PowerConsumption tool on run if it is not already built and setup the tool to be called. With these, a full E2E run is possible locally.
Add PowerConsumption to the scenario measurement.
Add Battery level and voltage specific measurements, along with joule usage calculations.