Add basic local testing script that builds and runs benchmarks (#3142)
Adds local script that builds and runs coreruns using BDN with support for CoreRun, Mono Interpreter, and Mono JIT with a placeholder for MonoAOTLLVM runtypes. CoreRun is consistently successful on both Windows and Linux while the Mono types are only consistently successful on Linux.

Started work on having a central artifact path separate from the runtime repo path. This will allow for all builds to be compared at the same time and save time by not having to rebuild everything everytime once functionality for use prebuilt is added. Added pre and post run kills for dotnet to ensure file handles are not held so user can cleanup the runs if wanted.


* Add osx as a supported OS

Co-authored-by: Milos Kotlar <kotlarmilos@gmail.com>

---------

Co-authored-by: Milos Kotlar <kotlarmilos@gmail.com>