Add strength reduction benchmarks
This adds strength reduction benchmarks for arrays of a few different
element sizes, motivated by the differences in codegen.
2: load
3: lea + load
4: load
8: load
12: lea + load
16: shl + load
29: imul

Each size has 3 variants of benchmarks: an array version, a span
version, and a fully strength reduced manual version. The JIT is
expected to be able to transform the array version into the strength
reduced version soon. The span version will also be transformed, but not
quite all the way (the strength reduction will not be able to fold in
the base byref of the span).