Update Maui version flows (#4321)
Updated how we get the dotnet version and rollback file for Maui .Net testing. Specifically updated from using maui generated information stored as json at certain urls to generating it based on maestro/darc updated version information now stored in the dotnet/performance Version.Details.XML. This means that we will be in charge of merging the maestro PRs for these versions as well as meaning we will have the option to further gate and better track versions between runs.