Fix Alpine builds and add try except (#4654)
For Alpine builds we were not respecting the osSubGroup variable
and as such were not using the correct rid when building. I also
added a try block around our call to run, as that can throw, and
we still want to fall back to the sas token for now if the cert
fails.