Fix WASM dotnet pathing bug (#4655)
* Fix WASM dotnet pathing bug

For WASM, we use a pre-downloaded version of dotnet, and do not install
it using `dotnet-install.ps1`. As a result we setup the environment using
the `dotnet_path` variable instead of `install_dir`. This change will
now set the path to `dotnet` with the same precedence as our setup
script does. Preferring `dotnet_path` if it exists, and falling back to
`install_dir` if it does not.

* Add join for executable

* Add brackets to if else statement