Fix for Mac Certs
From my work it does not appear that there is any programmatic way to recover
the public/private key pair in a Key Chain without having to enter a password.
As a result, we are going to grab the certificates directly off the disk, like
we do on Linux machines.